+++ web/core/src/main/java/org/seedstack/seed/web/internal/cors/WebCORSPlugin.java
@@ -15,6 +15,7 @@ import org.seedstack.seed.core.internal.AbstractSeedPlugin;
 import org.seedstack.seed.web.WebConfig;
 import org.seedstack.seed.web.spi.FilterDefinition;
 import org.seedstack.seed.web.spi.ListenerDefinition;
+import org.seedstack.seed.web.spi.SeedFilterPriority;
 import org.seedstack.seed.web.spi.ServletDefinition;
 import org.seedstack.seed.web.spi.WebProvider;
 import org.slf4j.Logger;
@@ -60,7 +61,7 @@ public class WebCORSPlugin extends AbstractSeedPlugin implements WebProvider {
             LOGGER.info("CORS support enabled on {}", corsConfig.getPath());
 
             FilterDefinition filterDefinition = new FilterDefinition("web-cors", CORSFilter.class);
-            filterDefinition.setPriority(1000);
+            filterDefinition.setPriority(SeedFilterPriority.CORS);
             filterDefinition.setAsyncSupported(true);
             filterDefinition.addInitParameters(buildCorsParameters());
             filterDefinition.addMappings(new FilterDefinition.Mapping(corsConfig.getPath()));