+++ web/core/src/main/java/org/seedstack/seed/web/internal/diagnostic/WebDiagnosticPlugin.java
@@ -16,6 +16,7 @@ import org.seedstack.seed.core.internal.AbstractSeedPlugin;
 import org.seedstack.seed.web.WebConfig;
 import org.seedstack.seed.web.spi.FilterDefinition;
 import org.seedstack.seed.web.spi.ListenerDefinition;
+import org.seedstack.seed.web.spi.SeedFilterPriority;
 import org.seedstack.seed.web.spi.ServletDefinition;
 import org.seedstack.seed.web.spi.WebProvider;
 import org.slf4j.Logger;
@@ -77,7 +78,7 @@ public class WebDiagnosticPlugin extends AbstractSeedPlugin implements WebProvid
             LOGGER.info("Per-request diagnostic enabled, a diagnostic file will be dumped for each request exception");
 
             FilterDefinition filterDefinition = new FilterDefinition("web-diagnostic", WebDiagnosticFilter.class);
-            filterDefinition.setPriority(10000);
+            filterDefinition.setPriority(SeedFilterPriority.DIAGNOSTIC);
             filterDefinition.setAsyncSupported(true);
             filterDefinition.addMappings(new FilterDefinition.Mapping("/*"));
             return Lists.newArrayList(filterDefinition);