+++ web/security/src/main/java/org/seedstack/seed/web/internal/security/WebSecurityModule.java
@@ -78,9 +78,8 @@ class WebSecurityModule extends ShiroWebModule {
 
         bind(WebConfig.SecurityConfig.class);
 
-        // Bind Seed filters
+        // Bind filters which are not PatchMatchingFilters
         bind(AntiXsrfFilter.class);
-        bind(X509CertificateFilter.class);
 
         // Bind custom filters not extending PathMatchingFilter as Shiro doesn't do it
         for (Class<? extends Filter> customFilter : customFilters) {